Commands, Configs, 'n Stuffs.
Adds commands and language strings for those.

Adds a config option for preventing fall damage. When I originally wrote the config file I expected the need for quite a few more options, but I've implemented most of them as region flags or permissions.

After a little more testing, I think this will be ready for the initial release after months of procrastination.